+++ src/components/header/bottom-header/right-menu/right-menu.module.css
@@ -10,7 +10,6 @@
     padding: 8px 14px;
     border-radius: 0;
   }
-
 }
 
 .input {
