+++ src/components/header/top-header/top-header.module.css
@@ -3,7 +3,6 @@
   flex-direction: row;
   justify-content: space-between;
   width: 1440px;
-  padding: 0 20px;
 }
 
 .topHeader {
