+++ src/global.module.css
@@ -53,12 +53,6 @@ body {
   min-height: 100vh;
 }
 
-.root {
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-}
-
 .heading {
   font-size: 2.5rem;
   font-weight: 700;
