+++ src/main.tsx
@@ -1,13 +1,11 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
-import styles from './global.module.css';
+import './global.module.css';
 import { RouterProvider } from 'react-router-dom';
 import { router } from './router/router';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <div className={styles.root}>
-      <RouterProvider router={router} />
-    </div>
+    <RouterProvider router={router} />
   </StrictMode>
 );
