+++ src/pages/not-found-page/not-found.module.css
@@ -44,5 +44,5 @@
 .backButton {
   font-size: 16px;
   font-weight: 400;
-  padding: 15px 30px
+  padding: 15px 30px;
 }
