+++ src/router/root.module.css
@@ -1,0 +1,5 @@
+.root {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+}
