+++ src/router/root.tsx
@@ -1,13 +1,16 @@
 import { Header } from '../components/header/header';
 import { Footer } from '../components/footer/footer';
-import { Section } from '../components/section/section';
+import { Section } from '../components/main-section/main-section';
+import styles from './root.module.css';
 
 export const Root = () => {
   return (
     <>
-      <Header />
-      <Section />
-      <Footer />
+      <div className={styles.root}>
+        <Header />
+        <Section />
+        <Footer />
+      </div>
     </>
   );
 };
